+++ lib/types/Config.ts
@@ -15,6 +15,7 @@ export default interface Config {
     treasury: {
       primary: string;
       secondary: string;
+      ingame: string;
     };
 
     backend: string;